Fix missing files on commit due to LF CRLF conflicts with git

+++ src/main/java/net/mcreator/radiant/procedures/ProgressionOnEffectActiveTickProcedure.java
@@ -16,7 +16,7 @@ public class ProgressionOnEffectActiveTickProcedure {
 				if (entity instanceof Player _player)
 					_player.giveExperiencePoints(-(1));
 				if (entity instanceof LivingEntity _entity)
-					_entity.setHealth((float) ((entity instanceof LivingEntity _livEnt ? _livEnt.getHealth() : -1) + 1));
+					_entity.setHealth((entity instanceof LivingEntity _livEnt ? _livEnt.getHealth() : -1) + 1);
 			}
 		} else {
 			if (entity instanceof LivingEntity _entity)
